+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -35,6 +35,7 @@ export type FileListObject = ClosedEnum<typeof FileListObject>;
 export const FileListPurpose = {
   Retrieval: "retrieval",
   KnowledgeDatasource: "knowledge_datasource",
+  Batch: "batch",
 } as const;
 /**
  * The intended purpose of the uploaded file.
@@ -195,7 +196,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-27T17:20:06.978Z",
+    "2025-02-28T09:18:03.386Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -229,7 +230,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-27T17:20:06.978Z"))
+  created: z.date().default(() => new Date("2025-02-28T09:18:03.386Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {