+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -21,6 +21,7 @@ export type FileT = {
 export const Purpose = {
   Retrieval: "retrieval",
   KnowledgeDatasource: "knowledge_datasource",
+  Batch: "batch",
 } as const;
 /**
  * The intended purpose of the uploaded file.
@@ -44,6 +45,7 @@ export type FileUploadRequestBody = {
 export const FileUploadPurpose = {
   Retrieval: "retrieval",
   KnowledgeDatasource: "knowledge_datasource",
+  Batch: "batch",
 } as const;
 /**
  * The intended purpose of the uploaded file.
@@ -245,7 +247,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-27T17:20:06.978Z",
+    "2025-02-28T09:18:03.386Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -279,7 +281,7 @@ export const FileUploadResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-27T17:20:06.978Z"))
+  created: z.date().default(() => new Date("2025-02-28T09:18:03.386Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {