+++ packages/orq-rc/RELEASES.md
@@ -609,3 +609,13 @@ Based on:
 - [typescript v3.2.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.21 - packages/orq-rc
+
+## 2025-02-28 09:19:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.507.0 (2.533.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.22 - packages/orq-rc